requirements.txt
numpy<2.0,>=1.16.0
scipy
matplotlib>=2.0.2
scikit-learn
tensorflow-gpu==2.0.0
setuptools>=41.0.0
pylint